+++ CommunityPatchExtras/Package/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+
+- Added World Save Notice Mode HUD setting (RelocateToTopLeft, Mute, Native) to prevent autosave notifications from obscuring the center of the screen.
 ## 0.1.1
 
 - Added icon